Add more questions to the first start setup
