+++ cloudnet/src/main/resources/lang/english.properties
@@ -7,7 +7,12 @@ cloudnet-init-eula=Do you agree to the Mojang EULA (https://account.mojang.com/d
 cloudnet-init-eula-not-accepted=If you don't accept the eula, you cannot run a minecraft server!
 cloudnet-init-config-low-memory-warning=The cloud has less than 2GB of memory available, perhaps it can only start a few services
 cloudnet-init-setup-node-id=What should the name of this CloudNet node be?
-cloudnet-init-setup-tasks-default-installation=Which tasks should be created by default?
+cloudnet-init-setup-tasks-should-install-proxy=Should a default proxy be created?
+cloudnet-init-setup-tasks-should-install-server=Should a default lobby be created?
+cloudnet-init-setup-tasks-proxy-environment=What should be the environment of the proxy?
+cloudnet-init-setup-tasks-proxy-version=Which ServiceVersion should be ran on the proxies?
+cloudnet-init-setup-tasks-server-environment=What should be the environment of the servers?
+cloudnet-init-setup-tasks-server-version=Which ServiceVersion should be ran on the servers?
 cloudnet-init-setup-memory=How much memory (in MB) can the services running on this CloudNet instance use?
 cloudnet-init-setup-internal-host=On which host and port should CloudNet run on?
 cloudnet-init-setup-web-host=On which host and port should CloudNet's WebServer run on?
