+++ cloudnet/src/main/resources/lang/german.properties
@@ -7,7 +7,12 @@ cloudnet-init-eula=Akzeptierst du die Mojang EULA (https://account.mojang.com/do
 cloudnet-init-eula-not-accepted=Ohne die EULA zu akzeptieren, kannst du keinen Minecraft Server hosten!
 cloudnet-init-config-low-memory-warning=Die Cloud hat weniger als 2GB RAM zur Verfügung, möglicherweise können nur wenige Services gestartet werden
 cloudnet-init-setup-node-id=Wie soll diese CloudNet Node heißen?
-cloudnet-init-setup-tasks-default-installation=Welche Standardeinrichtung soll auf CloudNet installiert werden?
+cloudnet-init-setup-tasks-should-install-proxy=Soll standardmäßig ein Proxy erstellt werden?
+cloudnet-init-setup-tasks-should-install-server=Soll standardmäßig eine Lobby erstellt werden?
+cloudnet-init-setup-tasks-proxy-environment=Welches Environment soll dieser Proxy haben?
+cloudnet-init-setup-tasks-proxy-version=Welche Service Version soll auf diesem Proxy laufen?
+cloudnet-init-setup-tasks-server-environment=Welches Environment sollen die Server haben?
+cloudnet-init-setup-tasks-server-version=Welche Service Version soll auf den Servern laufen?
 cloudnet-init-setup-memory=Wie viel MB RAM dürfen die Services in dieser CloudNet 3 Instanz maximal verwenden?
 cloudnet-init-setup-internal-host=Auf welchem Host und Port soll CloudNet laufen?
 cloudnet-init-setup-web-host=Auf welchem Host und Port soll CloudNets interner WebServer laufen?
